+++ packages/dds/tree/src/feature-libraries/modular-schema/modularChangeFormat.ts
@@ -22,6 +22,7 @@ export const EncodedChangeAtomId = Type.Union([
 	Type.Tuple([ChangesetLocalIdSchema, RevisionTagSchema]),
 	ChangesetLocalIdSchema,
 ]);
+export type EncodedChangeAtomId = Static<typeof EncodedChangeAtomId>;
 
 const EncodedValueChange = Type.Object(
 	{
