+++ packages/dds/tree/src/feature-libraries/optional-field/optionalFieldChangeFormatV0.ts
@@ -13,17 +13,21 @@ const noAdditionalProps: ObjectOptions = { additionalProperties: false };
 export const EncodedRegisterId = Type.Union([EncodedChangeAtomId, Type.Null()]);
 export type EncodedRegisterId = Static<typeof EncodedRegisterId>;
 
+export const EncodedBuild = Type.Tuple([EncodedChangeAtomId]);
+export type EncodedBuild = Static<typeof EncodedBuild>;
+
 export const EncodedOptionalChangeset = <Schema extends TSchema>(tNodeChange: Schema) =>
 	Type.Object(
 		{
-			// Moves between detached fields.
-			// These entries should not be interpreted as "applied one after the other", but rather as
-			// "applied simultaneously". As such, this list should not contain duplicated src or dst entries.
-			m: Type.Optional(Type.Array(EncodedMove)),
+			// Subtrees being created. They start as detached.
+			b: Type.Optional(Type.Array(EncodedBuild)),
+			// Subtrees being moved.
+			m: EncodedMoves,
 			// Nested changes
 			c: EncodedChildChanges(tNodeChange),
-			// How to replace the current value of the field.
-			r: Type.Optional(EncodedReplace),
+			// Reserved ID for detaching the subtree from the field if it were to be populated.
+			// Only specified when the field is empty.
+			d: Type.Optional(EncodedRegisterId),
 		},
 		noAdditionalProps,
 	);
@@ -42,18 +46,6 @@ const EncodedChildChanges = <Schema extends TSchema>(tNodeChange: Schema) =>
 // Otherwise the intention is to move whatever node happens to be in the source register.
 // These entries should not be interpreted as "applied one after the other", but rather as "applied simultaneously".
 // As such, changesets should not contain duplicated src or dst entries.
-const EncodedMove = Type.Tuple([EncodedChangeAtomId, EncodedChangeAtomId]);
-
-const EncodedReplace = Type.Object(
-	{
-		// Whether the field is empty in the input context of this change.
-		e: Type.Boolean(),
-		// The ID for the node to put in this field, or undefined if the field should be emptied.
-		// Will be "self" when the intention is to keep the current node in this field.
-		s: Type.Optional(EncodedRegisterId),
-		// An ID to associate with the node (if any) which is detached by this edit.
-		d: EncodedChangeAtomId,
-	},
-	noAdditionalProps,
+const EncodedMoves = Type.Optional(
+	Type.Array(Type.Tuple([EncodedRegisterId, EncodedRegisterId, Type.Optional(Type.Boolean())])),
 );
-export type EncodedReplace = Static<typeof EncodedReplace>;
